+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to the "prettier-vscode" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [Unreleased]
+- New setting `jsxBracketSameLine`.
 
 ## [0.7.0]
 - Removed `Prettier` action.